Rename API to utils
