+++ src/lib/index.ts
@@ -1,0 +1,2 @@
+export * as grammar from "./grammar.js";
+export * from "../parser/utils/parse-file.js";
