+++ src/parser/index.ts
@@ -1,2 +1,2 @@
 export { parse } from "./parser.js";
-export * from "./api/index.js";
+export * from "./utils/index.js";
