+++ src/parser/utils/parse-directory.ts
