+++ src/parser/utils/parse-module.ts
